requirements.txt
google-cloud-speech==2.19.0
google-cloud-texttospeech==2.14.1
openai==0.27.6
pydub==0.25.1